Refactor server main to make it simple to build with pulgins
